+++ src/idunno.AtProto/AtProtoHttpClient.cs
@@ -560,6 +560,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="service">The <see cref="Uri"/> of the service to call.</param>
     /// <param name="endpoint">The endpoint on the <paramref name="service"/> to call.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
     [RequiresDynamicCode("Use a Get overload which takes JsonSerializerOptions instead.")]
@@ -568,14 +569,16 @@ public class AtProtoHttpClient<TResult> where TResult : class
         Uri service,
         string endpoint,
         HttpClient httpClient,
+        TimeSpan? timeout = null,
         CancellationToken cancellationToken = default)
     {
         return await Get(
             service: service,
             endpoint: endpoint,
             credentials: null,
-            onCredentialsUpdated: null,
             httpClient: httpClient,
+            timeout: timeout,
+            onCredentialsUpdated: null,
             subscribedLabelers: null,
             cancellationToken: cancellationToken).ConfigureAwait(false);
     }
@@ -605,6 +608,39 @@ public class AtProtoHttpClient<TResult> where TResult : class
             onCredentialsUpdated: null,
             httpClient: httpClient,
             subscribedLabelers: null,
+            timeout: null,
+            jsonSerializerOptions: jsonSerializerOptions,
+            cancellationToken: cancellationToken).ConfigureAwait(false);
+    }
+
+    /// <summary>
+    /// Performs an unauthenticated GET request against the supplied <paramref name="service"/> and <paramref name="endpoint"/>.
+    /// </summary>
+    /// <param name="service">The <see cref="Uri"/> of the service to call.</param>
+    /// <param name="endpoint">The endpoint on the <paramref name="service"/> to call.</param>
+    /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
+    /// <param name="jsonSerializerOptions"><see cref="JsonSerializerOptions"/> to apply during deserialization.</param>
+    /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>The task object representing the asynchronous operation.</returns>
+    [RequiresUnreferencedCode("Make sure all required types are preserved in the jsonSerializerOptions parameter.")]
+    [RequiresDynamicCode("Make sure all the required types are preserved in the jsonSerializerOptions parameter.")]
+    public async Task<AtProtoHttpResult<TResult>> Get(
+        Uri service,
+        string endpoint,
+        HttpClient httpClient,
+        TimeSpan? timeout,
+        JsonSerializerOptions jsonSerializerOptions,
+        CancellationToken cancellationToken = default)
+    {
+        return await Get(
+            service: service,
+            endpoint: endpoint,
+            credentials: null,
+            onCredentialsUpdated: null,
+            httpClient: httpClient,
+            subscribedLabelers: null,
+            timeout: timeout,
             jsonSerializerOptions: jsonSerializerOptions,
             cancellationToken: cancellationToken).ConfigureAwait(false);
     }
@@ -616,6 +652,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="endpoint">The endpoint on the <paramref name="service"/> to call.</param>
     /// <param name="credentials">The <see cref="AtProtoCredential"/> to use when calling <paramref name="service"/>.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
     /// <param name="requestHeaders">A collection of HTTP headers to send with the request.</param>
     /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
@@ -630,6 +667,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         string endpoint,
         AtProtoCredential? credentials,
         HttpClient httpClient,
+        TimeSpan? timeout = null,
         Action<AtProtoCredential>? onCredentialsUpdated = null,
         ICollection<NameValueHeaderValue>? requestHeaders = null,
         IEnumerable<Did>? subscribedLabelers = null,
@@ -649,6 +687,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             credentials: credentials,
             httpClient: httpClient,
             retry: true,
+            timeout: timeout,
             onCredentialsUpdated: onCredentialsUpdated,
             subscribedLabelers: subscribedLabelers,
             jsonSerializerOptions: _jsonSerializationOptionsDefault,
@@ -663,6 +702,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="credentials">The <see cref="AtProtoCredential"/> to use when calling <paramref name="service"/>.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
     /// <param name="jsonSerializerOptions"><see cref="JsonSerializerOptions"/> to apply during deserialization.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
     /// <param name="requestHeaders">A collection of HTTP headers to send with the request.</param>
     /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
@@ -678,6 +718,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         AtProtoCredential? credentials,
         HttpClient httpClient,
         JsonSerializerOptions jsonSerializerOptions,
+        TimeSpan? timeout = null,
         Action<AtProtoCredential>? onCredentialsUpdated = null,
         ICollection<NameValueHeaderValue>? requestHeaders = null,
         IEnumerable<Did>? subscribedLabelers = null,
@@ -698,6 +739,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             credentials: credentials,
             httpClient: httpClient,
             retry: true,
+            timeout: timeout,
             onCredentialsUpdated: onCredentialsUpdated,
             subscribedLabelers: subscribedLabelers,
             jsonSerializerOptions: jsonSerializerOptions,
@@ -711,6 +753,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="endpoint">The endpoint on the <paramref name="service"/> to call.</param>
     /// <param name="credentials">The <see cref="AtProtoCredential"/> to use when calling <paramref name="service"/>.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the provided have been updated during the HTTP POST.</param>
     /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
@@ -722,6 +765,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         string endpoint,
         AtProtoCredential? credentials,
         HttpClient httpClient,
+        TimeSpan? timeout = null,
         Action<AtProtoCredential>? onCredentialsUpdated = null,
         IEnumerable<Did>? subscribedLabelers = null,
         CancellationToken cancellationToken = default)
@@ -733,6 +777,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             requestHeaders: _extraRequestHeaders,
             credentials: credentials,
             httpClient: httpClient,
+            timeout: timeout,
             onCredentialsUpdated: onCredentialsUpdated,
             subscribedLabelers: subscribedLabelers,
             cancellationToken: cancellationToken).ConfigureAwait(false);
@@ -746,6 +791,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="credentials">The <see cref="AtProtoCredential"/> to use when calling <paramref name="service"/>.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
     /// <param name="jsonSerializerOptions"><see cref="JsonSerializerOptions"/> to apply during deserialization.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the provided have been updated during the HTTP POST.</param>
     /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
@@ -758,6 +804,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         AtProtoCredential? credentials,
         HttpClient httpClient,
         JsonSerializerOptions jsonSerializerOptions,
+        TimeSpan? timeout = null,
         Action<AtProtoCredential>? onCredentialsUpdated = null,
         IEnumerable<Did>? subscribedLabelers = null,
         CancellationToken cancellationToken = default)
@@ -769,6 +816,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             requestHeaders: _extraRequestHeaders,
             credentials: credentials,
             httpClient: httpClient,
+            timeout: timeout,
             onCredentialsUpdated: onCredentialsUpdated,
             subscribedLabelers: subscribedLabelers,
             jsonSerializerOptions: jsonSerializerOptions,
@@ -783,6 +831,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="endpoint">The endpoint on the <paramref name="service"/> to call.</param>
     /// <param name="record">An optional object to serialize to JSON and send as the request body.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
@@ -793,6 +842,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         string endpoint,
         TRecord? record,
         HttpClient httpClient,
+        TimeSpan? timeout = null,
         IEnumerable<Did>? subscribedLabelers = null,
         CancellationToken cancellationToken = default)
     {
@@ -803,6 +853,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             requestHeaders: null,
             credentials: null,
             httpClient: httpClient,
+            timeout: timeout,
             onCredentialsUpdated: null,
             subscribedLabelers: subscribedLabelers,
             cancellationToken: cancellationToken).ConfigureAwait(false);
@@ -817,6 +868,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="record">An optional object to serialize to JSON and send as the request body.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
     /// <param name="jsonSerializerOptions"><see cref="JsonSerializerOptions"/> to apply during deserialization.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
@@ -828,6 +880,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         TRecord? record,
         HttpClient httpClient,
         JsonSerializerOptions jsonSerializerOptions,
+        TimeSpan? timeout = null,
         IEnumerable<Did>? subscribedLabelers = null,
         CancellationToken cancellationToken = default)
     {
@@ -838,6 +891,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             requestHeaders: null,
             credentials: null,
             httpClient: httpClient,
+            timeout: timeout,
             onCredentialsUpdated: null,
             subscribedLabelers: subscribedLabelers,
             jsonSerializerOptions: jsonSerializerOptions,
@@ -853,6 +907,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="record">An optional object to serialize to JSON and send as the request body.</param>
     /// <param name="credentials">The <see cref="AtProtoCredential"/> to use when calling <paramref name="service"/>.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the provided have been updated during the HTTP POST.</param>
     /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
@@ -865,6 +920,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         TRecord? record,
         AtProtoCredential credentials,
         HttpClient httpClient,
+        TimeSpan? timeout = null,
         Action<AtProtoCredential>? onCredentialsUpdated = null,
         IEnumerable<Did>? subscribedLabelers = null,
         CancellationToken cancellationToken = default)
@@ -876,6 +932,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             requestHeaders: null,
             credentials: credentials,
             httpClient: httpClient,
+            timeout: timeout,
             onCredentialsUpdated: onCredentialsUpdated,
             subscribedLabelers: subscribedLabelers,
             cancellationToken: cancellationToken).ConfigureAwait(false);
@@ -891,6 +948,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="credentials">The <see cref="AtProtoCredential"/> to use when calling <paramref name="service"/>.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
     /// <param name="jsonSerializerOptions"><see cref="JsonSerializerOptions"/> to apply during deserialization.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the provided have been updated during the HTTP POST.</param>
     /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
@@ -904,6 +962,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         AtProtoCredential credentials,
         HttpClient httpClient,
         JsonSerializerOptions jsonSerializerOptions,
+        TimeSpan? timeout = null,
         Action<AtProtoCredential>? onCredentialsUpdated = null,
         IEnumerable<Did>? subscribedLabelers = null,
         CancellationToken cancellationToken = default)
@@ -915,6 +974,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             requestHeaders: null,
             credentials: credentials,
             httpClient: httpClient,
+            timeout: timeout,
             onCredentialsUpdated: onCredentialsUpdated,
             subscribedLabelers: subscribedLabelers,
             jsonSerializerOptions: jsonSerializerOptions,
@@ -931,6 +991,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="requestHeaders">A collection of HTTP headers to send with the request.</param>
     /// <param name="credentials">The <see cref="AtProtoCredential"/> to use when calling <paramref name="service"/>.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the provided have been updated during the HTTP POST.</param>
     /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
@@ -946,6 +1007,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         ICollection<NameValueHeaderValue>? requestHeaders,
         AtProtoCredential? credentials,
         HttpClient httpClient,
+        TimeSpan? timeout = null,
         Action<AtProtoCredential>? onCredentialsUpdated = null,
         IEnumerable<Did>? subscribedLabelers = null,
         CancellationToken cancellationToken = default)
@@ -964,6 +1026,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             credentials: credentials,
             httpClient: httpClient,
             retry: true,
+            timeout: timeout,
             onCredentialsUpdated: onCredentialsUpdated,
             subscribedLabelers: subscribedLabelers,
             jsonSerializerOptions: _jsonSerializationOptionsDefault,
@@ -981,6 +1044,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="credentials">The <see cref="AtProtoCredential"/> to use when calling <paramref name="service"/>.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
     /// <param name="jsonSerializerOptions"><see cref="JsonSerializerOptions"/> to apply during deserialization.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the provided have been updated during the HTTP POST.</param>
     /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
@@ -997,6 +1061,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         AtProtoCredential? credentials,
         HttpClient httpClient,
         JsonSerializerOptions jsonSerializerOptions,
+        TimeSpan? timeout = null,
         Action<AtProtoCredential>? onCredentialsUpdated = null,
         IEnumerable<Did>? subscribedLabelers = null,
         CancellationToken cancellationToken = default)
@@ -1016,6 +1081,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             credentials: credentials,
             httpClient: httpClient,
             retry: true,
+            timeout: timeout,
             onCredentialsUpdated: onCredentialsUpdated,
             subscribedLabelers: subscribedLabelers,
             jsonSerializerOptions: jsonSerializerOptions,
@@ -1032,6 +1098,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="contentHeaders">A collection of HTTP content headers to send with the request content.</param>
     /// <param name="credentials">The <see cref="AtProtoCredential"/> to use when calling <paramref name="service"/>.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
     /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
@@ -1047,6 +1114,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         ICollection<NameValueHeaderValue>? contentHeaders,
         AtProtoCredential credentials,
         HttpClient httpClient,
+        TimeSpan? timeout = null,
         Action<AtProtoCredential>? onCredentialsUpdated = null,
         CancellationToken cancellationToken = default)
     {
@@ -1077,6 +1145,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             credentials: credentials,
             httpClient: httpClient,
             retry: true,
+            timeout: timeout,
             onCredentialsUpdated: onCredentialsUpdated,
             subscribedLabelers: null,
             jsonSerializerOptions: _jsonSerializationOptionsDefault,
@@ -1094,6 +1163,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="credentials">The <see cref="AtProtoCredential"/> to use when calling <paramref name="service"/>.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
     /// <param name="jsonSerializerOptions"><see cref="JsonSerializerOptions"/> to apply during deserialization.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
     /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
@@ -1110,6 +1180,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         AtProtoCredential credentials,
         HttpClient httpClient,
         JsonSerializerOptions jsonSerializerOptions,
+        TimeSpan? timeout = null,
         Action<AtProtoCredential>? onCredentialsUpdated = null,
         CancellationToken cancellationToken = default)
     {
@@ -1141,6 +1212,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             credentials: credentials,
             httpClient: httpClient,
             retry: true,
+            timeout: timeout,
             onCredentialsUpdated: onCredentialsUpdated,
             subscribedLabelers: null,
             jsonSerializerOptions: jsonSerializerOptions,
@@ -1158,6 +1230,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
     /// <param name="credentials">The <see cref="AtProtoCredential"/> to use when calling <paramref name="service"/>.</param>
     /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
     /// <param name="jsonSerializerOptions"><see cref="JsonSerializerOptions"/> to apply during deserialization.</param>
+    /// <param name="timeout">An optional <see cref="TimeSpan"/> to apply as the request timeout.</param>
     /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
     /// <param name="cancellationToken">An optional cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
@@ -1175,6 +1248,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         AtProtoCredential credentials,
         HttpClient httpClient,
         JsonSerializerOptions jsonSerializerOptions,
+        TimeSpan? timeout = null,
         Action<AtProtoCredential>? onCredentialsUpdated = null,
         CancellationToken cancellationToken = default)
     {
@@ -1207,6 +1281,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
             retry: true,
             onCredentialsUpdated: onCredentialsUpdated,
             subscribedLabelers: null,
+            timeout: timeout,
             jsonSerializerOptions: jsonSerializerOptions,
             cancellationToken: cancellationToken).ConfigureAwait(false);
     }
@@ -1331,10 +1406,6 @@ public class AtProtoHttpClient<TResult> where TResult : class
             httpRequestMessage.Headers.Add(header.Key, header.Value);
         }
 
-        // Force the response to be json.
-        httpRequestMessage.Headers.Accept.Clear();
-        httpRequestMessage.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue(MediaTypeNames.Application.Json));
-
         if (subscribedLabelers is not null)
         {
             List<string> labelerIdentifiers = [.. subscribedLabelers];
@@ -1351,6 +1422,12 @@ public class AtProtoHttpClient<TResult> where TResult : class
             {
                 httpRequestMessage.Headers.TryAddWithoutValidation(headerValue.Name, headerValue.Value);
             }
+        }
+
+        // Force the response to be json.
+        if (!httpRequestMessage.Headers.Contains("accept"))
+        {
+            httpRequestMessage.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue(MediaTypeNames.Application.Json));
         }
     }
 
@@ -1447,6 +1524,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
         ICollection<NameValueHeaderValue>? contentHeaders,
         AtProtoCredential? credentials,
         HttpClient httpClient,
+        TimeSpan? timeout,
         bool retry = false,
         Action<AtProtoCredential>? onCredentialsUpdated = null,
         IEnumerable<Did>? subscribedLabelers = null,
@@ -1511,6 +1589,11 @@ public class AtProtoHttpClient<TResult> where TResult : class
                 if (OnSendingRequest is not null)
                 {
                     await OnSendingRequest(httpRequestMessage, cancellationToken).ConfigureAwait(false);
+                }
+
+                if (timeout.HasValue)
+                {
+                    httpClient.Timeout = timeout.Value;
                 }
 
                 try
@@ -1688,6 +1771,7 @@ public class AtProtoHttpClient<TResult> where TResult : class
                                             credentials: credentials,
                                             httpClient: httpClient,
                                             retry: false,
+                                            timeout: timeout,
                                             onCredentialsUpdated: onCredentialsUpdated,
                                             subscribedLabelers: subscribedLabelers,
                                             jsonSerializerOptions: jsonSerializerOptions,